Migrate intr_ctrl into top_chip_system
The placeholder interrupt controller was outside of top_chip_system
originally because the intention was that a different implementation
be provided.

Standardising on rv_plic means that it is more sensible to move
the interrupt controller into the design to be consistent with the
other top levels, and reduces the potential for divergence between
top_chip_asic and top_chip_verilator.