Import of test_runner/plic_tests.hh from Sonata
Import a pared back version of the RV PLIC testing from
the Sonata project for the initial bring up of the RV PLIC
in Sunburst chip.

Note that it has been necessary to place the code and data
within the SRAM because the ROM is (i) not writable, causing
problems with global data, and (ii) somewhat too small for
the PLIC testing and logging code.

Extend the PLIC testing to include pattgen, the SPI hosts
and USBDEV.