[dv] Add virtual sequencer to top-level testbench
The base test starts a virtual sequence on this new virtual sequencer
that's specified by the UVM_TEST_SEQ plusarg that can be specified in
dvsim configuration files. Checking for the simulation exit condition
has been refactored so it is handled in the base virtual sequence rather
than in the test itself.

An example virtual sequence is provided that checks that outputs on GPIO
match the pattern produced by the 'chip_check' check program.