[dv,sw] Add (un-randomised) pattgen chip-level test
Add a pattgen vseq (based on the OpenTitan one) and sketch of a
program to test it. Put both of them together in dvsim as
`top_pattgen_smoke`.

As part of this, add peri_clk_if to top_chip_dv_if_bundle
so that the virtual sequence can use the correctly
synchronised reset.

Note that `top_pattgen_smoke` currently only 'passes' because a
constraint has been added that locks the pattgen parameters
to those currently hard-coded in the software part of the test.
To use randomised pattgen parameters, the virtual sequence
requires being able to pass the pattgen parameters to use to the SW,
but the mechanism to do this has not yet been ported/implemented.