[dv] Initial SW test-status & logger DV ROM hookup
Initial attempt at integrating the DV components of the software
test-status and fast-logging mechanisms.

Connect them to the ROM TL-UL interface for now, as the ROM doesn't need
to be written to and the status and logging interfaces only need writes.
Note that this has required a tweak to the ROM TL-UL adapter to
disable write errors using parameters.
Eventually, connect to the (presently non-existant) ibex core wrapper,
as in OpenTitan.