Add usbdev_aon_wake module for suspend/resume support
The usbdev_aon_wake module takes over the driving of the pullup
and the monitoring of the USB when the SoC is in a low power
state. This is only supported when the USB host controller has
put the USB device into the Suspend state.

Whilst the module is active it will detect disconnections,
bus resets and activity on the bus (which should be Resume
signalling). In response it issues a wakeup request to the power
management.

Software must be able to return control of the USB to the USB
device (having reinitialised its state if necessary) in under
20ms for the Resume operation to be successful.