[sw] Inline `irq_global_ctrl()`
This is a workaround for CHERIoT using interrupt-disabling backward
sentries to return from `irq_global_ctrl()`, thus removing the ability
to enable interrupts for more than a few cycles at a time.
Inlining the function removes the jump-and-link and associated
interrupt-disabling backward sentry, allowing the enabling
of interrupts to persist in the context of the caller and any
inheriting functions.

See https://github.com/CHERIoT-Platform/llvm-project/issues/120