Bring up of sw_logger_if faster DV logging
A number of the T-L tests rely upon the sw_logger_if either
for the correct timing (faster than simulating a UART) or
because the logging messages are used as a synchronisation
mechanism between the DV sequences and the device-side test
software.

Make this mechanism CHERIoT capability pointer-aware, which
unfortunately requires keeping and parsing the format specifier
string to some degree. Previously the internal DV logging code
assumed that all arguments were uint32_t quantities but this
is no longer the case.

Note that capability relocations persist in the emitted binary
for the global static fields in the `.logs.fields` INFO section,
even though that section has been dropped, so the bootstrap
software must skip performing those relocations.